+++ src/ValidationPackage/CoreValidationPackage.php
@@ -27,6 +27,7 @@ class CoreValidationPackage implements ValidationPackage
             'float' => $p.'typeFloat',
             'double' => $p.'typeDouble',
             'digits' => $p.'digits',
+            'date' => $p.'date',
 
             'in' => $p.'inArray',
             'all' => $p.'forAll',
@@ -50,7 +51,8 @@ class CoreValidationPackage implements ValidationPackage
             'float' => 'The {{attribute}} is not a valid float.',
             'double' => 'The {{attribute}} is not a valid double.',
             'digits' => 'The {{attribute}} must only contain digits.',
-            'in' => 'The {{attribute}} is must be one of: {{accepted}}.',
+            'date' => 'The {{attribute}} must be a valid date.',
+            'in' => 'The {{attribute}} must be one of: {{accepted}}.',
             'regex' => 'The {{attribute}} does not match the pattern: {{pattern}}.',
             'email' => 'The {{attribute}} is not a valid email.',
             'regex_exclude' => 'The {{attribute}} must not match the pattern: {{pattern}}',